int64_t type chnaged to int_ptr_t

+++ tests/test.c
@@ -28,9 +28,10 @@ bool trivial_equal ( llist_node node1, llist_node node2 )
 {
     return ( node1 == node2 );
 }
+
 int trivial_comperator ( llist_node first, llist_node second )
 {
-    return ( ( int64_t ) first ) - ( ( int64_t ) second );
+    return ( ( intptr_t )first  - (intptr_t)second );
 }
 
 void trivial_node_func ( llist_node node )
@@ -783,7 +784,7 @@ END_TEST
 
 void * list_put_data(void * arg)
 {
-    int64_t i = 50;
+    ptrdiff_t i = 50;
     int retval;
     ck_assert ( arg != NULL );
     llist listToTest = (llist)arg;
@@ -811,6 +812,7 @@ void * list_get_data(void * arg)
         //wait for 100 milliseconds
         printf("find attempted");
         usleep(1000*100);
+        //printf ("the address to jump %p", &&start_waiting);
         goto start_waiting;
     }
     return NULL;
